+++ lib/presentation/onboarding/screens/onboarding_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:on_time_front/core/di/di_setup.dart';
-import 'package:on_time_front/presentation/onboarding/preparation_order/screens/preparation_reorderable_list.dart';
+import 'package:on_time_front/presentation/onboarding/preparation_order/screens/preparation_order_form.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_name_select/cubit/preparation_name/preparation_name_cubit.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_name_select/screens/preparation_name_form.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_time/cubit/preparation_time_cubit.dart';
@@ -105,7 +105,7 @@ class _OnboardingFormState extends State<OnboardingForm>
                     onPageChanged: _handlePageViewChanged,
                     children: <Widget>[
                       PreparationNameForm(),
-                      PreparationReorderField(),
+                      PreparationOrderForm(),
                       PreparationTimeField(),
                       ScheduleSpareTimeForm(),
                     ],
